requirements.txt
APScheduler==3.8.1
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
Deprecated==1.2.13
h11==0.12.0
idna==3.3
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
Pillow==9.0.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.0
requests==2.26.0
requests-oauthlib==1.3.0
rq==1.10.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
tweepy==4.4.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
wget==3.2
wrapt==1.13.3
wsproto==1.0.0
